Still Broken
Not entirely sure what is going on with my code, but Ill figure it out eventually.

Trying to show/hide the content divs is just showing the space that the divs take up. All of the actual content is invisible, even though the active div's content is not listed as hidden or display "none" in the Inspector.

Trying to add/remove an explicit hide and show css class is not working either. I had better luck with the wonky behaviour of my first attempt.