It now Kind of works
This is working, but wierdly. You have to click on one button to trigger HideLanding();
Then you must click on a DIFFERENT button to start toggling content.
Not sure what to do here, perhaps use a different event listener for Landing such that if you click anywhere on Landing
It hides it and have redwolf default to shown on page.